First Comment
Solutions for:
Display the location for each move in the format (col, row) in the move history list.
Bold the currently selected item in the move list.
Rewrite Board to use two loops to make the squares instead of hardcoding them.
Add a toggle button that lets you sort the moves in either ascending or descending order.
When no one wins, display a message about the result being a draw.

New feature:
Square grid matrix scaling. Assign a new number of rows/cols and the interface figures out the winning combinations for the new grid size.